requirements.txt
pymupdf==1.23.8
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
pandas==2.1.3
ollama==0.1.7
